Now uses ObjectStream instead of Iterator and DataStream
git-svn-id: https://svn.apache.org/repos/asf/incubator/opennlp/tags/asf_migration@1048327 13f79535-47bb-0310-9956-ffa450edef68

+++ opennlp-tools/trunk/src/main/java/opennlp/tools/util/Mean.java
@@ -52,7 +52,7 @@ public class Mean {
    * @param count number of times the value should be added to
    * arithmetic mean.
    */
-  public void add(double value, int count) {
+  public void add(double value, long count) {
     sum += value * count;
     this.count += count;    
   }
